--- cigaradvisor/blocks/article-recommendations/article-recommendations.css
@@ -1,6 +1,0 @@
-.article-recommendations.block .recommendations-heading {
-    font-family: var(--ff-opensans);
-    font-weight: var(--font-weight-bold);
-    font-size: var(--heading-font-size-xxs);
-    line-height: var(--line-height-m);
-}
