+++ cigaradvisor/styles/templates/blog-post.css
@@ -78,6 +78,13 @@ body.blog-post-template main .default-content-wrapper h3 {
   margin: 2rem 0 1rem;
 }
 
+body.blog-post-template main .recommendations-heading {
+  font-family: var(--ff-opensans);
+  font-weight: var(--font-weight-bold);
+  font-size: var(--heading-font-size-xxs);
+  line-height: var(--line-height-m);
+}
+
 @media screen and (min-width: 1200px) {
   .blog-post-template main .default-content-wrapper {
     padding: 0;
